Added column width and resizing, and improved the handling of numberOfRows

+++ MBTableGridContentView.m
@@ -37,6 +37,7 @@
 - (void)_setObjectValue:(id)value forColumn:(NSUInteger)columnIndex row:(NSUInteger)rowIndex;
 - (BOOL)_canEditCellAtColumn:(NSUInteger)columnIndex row:(NSUInteger)rowIndex;
 - (void)_setStickyColumn:(MBTableGridEdge)stickyColumn row:(MBTableGridEdge)stickyRow;
+- (float)_widthForColumn:(NSUInteger)columnIndex;
 - (id)_backgroundColorForColumn:(NSUInteger)columnIndex row:(NSUInteger)rowIndex;
 - (MBTableGridEdge)_stickyColumn;
 - (MBTableGridEdge)_stickyRow;
@@ -92,9 +93,10 @@
 
 
 - (void)drawRect:(NSRect)rect
-{	
-	NSUInteger numberOfColumns = [[self tableGrid] numberOfColumns];
-	NSUInteger numberOfRows = [[self tableGrid] numberOfRows];
+{
+    
+	NSUInteger numberOfColumns = [self tableGrid].numberOfColumns;
+	NSUInteger numberOfRows = [self tableGrid].numberOfRows;
 	
 	NSUInteger firstColumn = NSNotFound;
 	NSUInteger lastColumn = numberOfColumns - 1;
@@ -132,7 +134,6 @@
 		row = firstRow;
 		while (row <= lastRow) {
 			NSRect cellFrame = [self frameOfCellAtColumn:column row:row];
-			
 			// Only draw the cell if we need to
 			if ([self needsToDrawRect:cellFrame]) {
                 
@@ -152,7 +153,7 @@
 	NSIndexSet *selectedColumns = [[self tableGrid] selectedColumnIndexes];
 	NSIndexSet *selectedRows = [[self tableGrid] selectedRowIndexes];
 	
-	if([selectedColumns count] && [selectedRows count] && [[self tableGrid] numberOfColumns] > 0 && [[self tableGrid] numberOfRows] > 0) {
+	if([selectedColumns count] && [selectedRows count] && [self tableGrid].numberOfColumns > 0 && [self tableGrid].numberOfRows > 0) {
 		NSRect selectionTopLeft = [self frameOfCellAtColumn:[selectedColumns firstIndex] row:[selectedRows firstIndex]];
 		NSRect selectionBottomRight = [self frameOfCellAtColumn:[selectedColumns lastIndex] row:[selectedRows lastIndex]];
 		
@@ -194,9 +195,9 @@
 	}
 	
 	// Draw the column drop indicator
-	if (isDraggingColumnOrRow && dropColumn != NSNotFound && dropColumn <= [[self tableGrid] numberOfColumns] && dropRow == NSNotFound) {
+	if (isDraggingColumnOrRow && dropColumn != NSNotFound && dropColumn <= [self tableGrid].numberOfColumns && dropRow == NSNotFound) {
 		NSRect columnBorder;
-		if(dropColumn < [[self tableGrid] numberOfColumns]) {
+		if(dropColumn < [self tableGrid].numberOfColumns) {
 			columnBorder = [self rectOfColumn:dropColumn];
 		} else {
 			columnBorder = [self rectOfColumn:dropColumn-1];
@@ -215,9 +216,9 @@
 	}
 	
 	// Draw the row drop indicator
-	if (isDraggingColumnOrRow && dropRow != NSNotFound && dropRow <= [[self tableGrid] numberOfRows] && dropColumn == NSNotFound) {
+	if (isDraggingColumnOrRow && dropRow != NSNotFound && dropRow <= [self tableGrid].numberOfRows && dropColumn == NSNotFound) {
 		NSRect rowBorder;
-		if(dropRow < [[self tableGrid] numberOfRows]) {
+		if(dropRow < [self tableGrid].numberOfRows) {
 			rowBorder = [self rectOfRow:dropRow];
 		} else {
 			rowBorder = [self rectOfRow:dropRow-1];
@@ -236,7 +237,7 @@
 	}
 	
 	// Draw the cell drop indicator
-	if (!isDraggingColumnOrRow && dropRow != NSNotFound && dropRow <= [[self tableGrid] numberOfRows] && dropColumn != NSNotFound && dropColumn <= [[self tableGrid] numberOfColumns]) {
+	if (!isDraggingColumnOrRow && dropRow != NSNotFound && dropRow <= [self tableGrid].numberOfRows && dropColumn != NSNotFound && dropColumn <= [self tableGrid].numberOfColumns) {
 		NSRect cellFrame = [self frameOfCellAtColumn:dropColumn row:dropRow];
 		cellFrame.origin.x -= 2.0;
 		cellFrame.origin.y -= 2.0;
@@ -387,7 +388,7 @@
 
 - (void)resetCursorRects
 {
-    NSLog(@"%s - %f %f %f %f", __func__, grabHandleRect.origin.x, grabHandleRect.origin.y, grabHandleRect.size.width, grabHandleRect.size.height);
+    //NSLog(@"%s - %f %f %f %f", __func__, grabHandleRect.origin.x, grabHandleRect.origin.y, grabHandleRect.size.width, grabHandleRect.size.height);
 	// The main cursor should be the cell selection cursor
 	[self addCursorRect:[self visibleRect] cursor:[self _cellSelectionCursor]];
     [self addCursorRect:grabHandleRect cursor:[self _cellExtendSelectionCursor]];
@@ -508,12 +509,15 @@
 
 - (NSRect)rectOfColumn:(NSUInteger)columnIndex
 {
-	NSRect rect = NSMakeRect(0, 0, 60, [self frame].size.height);
+    
+    float width = [[self tableGrid] _widthForColumn:columnIndex];
+
+	NSRect rect = NSMakeRect(0, 0, width, [self frame].size.height);
 	//rect.origin.x += 60.0 * columnIndex;
 	
 	NSUInteger i = 0;
 	while(i < columnIndex) {
-		float headerWidth = rect.size.width;
+        float headerWidth = [[self tableGrid] _widthForColumn:i];
 		rect.origin.x += headerWidth;
 		i++;
 	}
@@ -523,6 +527,7 @@
 
 - (NSRect)rectOfRow:(NSUInteger)rowIndex
 {
+    
 	float heightForRow = 20.0;
 	NSRect rect = NSMakeRect(0, 0, [self frame].size.width, heightForRow);
 	
@@ -548,7 +553,7 @@
 - (NSInteger)columnAtPoint:(NSPoint)aPoint
 {
 	NSInteger column = 0;
-	while(column < [[self tableGrid] numberOfColumns]) {
+	while(column < [self tableGrid].numberOfColumns) {
 		NSRect columnFrame = [self rectOfColumn:column];
 		if(NSPointInRect(aPoint, columnFrame)) {
 			return column;
@@ -561,7 +566,7 @@
 - (NSInteger)rowAtPoint:(NSPoint)aPoint
 {
 	NSInteger row = 0;
-	while(row < [[self tableGrid] numberOfRows]) {
+	while(row < [self tableGrid].numberOfRows) {
 		NSRect rowFrame = [self rectOfRow:row];
 		if(NSPointInRect(aPoint, rowFrame)) {
 			return row;
